Sleep Test works with temp sensing
Sleeps for 32sec (8sec*4) using watchdog timer. Reads temp measurement
accurately and sends that and the ID to the wall unit, then goes back to
sleep.

Setting all pins to inputs initially causes the temp sensor to read
20deg higher approx, and when i attempted to set TX and power to inputs
it caused all sorts of problems. it is a recommended thing to do in
order to save power. Next step will be to switch the XBee on and off.