Fixed loads of shit
Changed temp buttons to interrupts using the rising edge. Changed
battstatus to string. changed the jsonPublish and dbPublish temp values
to strings, so we can pass a "--" when the remote sensor hasnt passed
data in a while.
Both json and db publishes function.
Buttons function.
App functions (wall temp was not working, but does now)